Add $options to items passed in Honeybadger.notify context.
Seems like a good compromise for including most of the relevant
component context without blowing up the GET request Honeybadger-js
makes. There's a chance it might be more likely to blow up on a real
request, but I think we're going to change Honeybadger-js to use POST so
it may be moot.